--- a/crates/nu_plugin_selector/src/selector.rs
+++ b/crates/nu_plugin_selector/src/selector.rs
@@ -139,8 +139,8 @@ fn retrieve_table(mut table: Table, columns: &Value) -> Vec<Value> {
                     .get(col)
                     .unwrap_or(&format!("Missing column: '{}'", &col))
                     .to_string();
-                if at_least_one_row_filled == false
-                    && val != String::from(format!("Missing column: '{}'", &col))
+                if !at_least_one_row_filled
+                    && val != format!("Missing column: '{}'", &col)
                 {
                     at_least_one_row_filled = true;
                 }