@fdncred 2021-09-25
cool! anxious to play with this. looking good.