@fdncred 2021-09-26
Hey @luccasmmg, if you have the will power for more, take a look at each table in the excel Wikipedia link. some are correct, some are not correct. Maybe there's a better way to get all the tables? what do you think? I knew this would be a tough task when I started playing with this because there are so many variations.